requirements.txt
alembic
email-to
fsspec
gcsfs
jsonschema
pandas
pg8000
pysftp
python-dotenv
requests
flask-shell-ipython

# [ML]
pipeline
scikit-learn
pickle-mixin
joblib
dill

# [framework]
Flask
Flask-APScheduler
Flask-Marshmallow
Marshmallow-SQLAlchemy
Flask-Migrate
Flask-RESTful
Flask-SQLAlchemy
Flask_WTF

# [server]
Gunicorn
# uWSGI